# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ninho API
  version: 1.4.0
  description: API residencial versionada, autenticada por cookies HttpOnly e segregada por RBAC.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      operationId: login
      responses:
        '200': { description: Sessão criada }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password/forgot:
    post:
      operationId: requestPasswordReset
      description: Sempre responde de forma genérica para impedir enumeração de contas.
      responses:
        '202': { description: Solicitação recebida }
  /auth/password/reset:
    post:
      operationId: resetPassword
      responses:
        '204': { description: Senha alterada e sessões anteriores invalidadas }
        '400': { description: Token inválido, expirado, usado ou senha fora da política }
  /homes:
    get:
      operationId: listHomes
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Residências autorizadas }
    post:
      operationId: createHome
      security: [{ cookieAuth: [] }]
      responses:
        '201': { description: Residência criada com piso e ambientes iniciais }
  /homes/{homeId}:
    patch:
      operationId: updateHome
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Residência atualizada e versão incrementada }
        '409': { $ref: '#/components/responses/VersionConflict' }
    delete:
      operationId: deleteHome
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: Residência excluída }
        '409': { description: Versão desatualizada ou última residência do proprietário }
  /homes/{homeId}/floors:
    get:
      operationId: listFloors
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses: { '200': { description: Pisos da residência } }
    post:
      operationId: createFloor
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses: { '201': { description: Piso criado } }
  /homes/{homeId}/floors/{floorId}:
    patch:
      operationId: updateFloor
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/FloorId' }
      responses:
        '200': { description: Piso atualizado e versão incrementada }
        '409': { $ref: '#/components/responses/VersionConflict' }
    delete:
      operationId: deleteFloor
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/FloorId' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: Piso excluído }
        '409': { description: Piso ocupado, último piso ou versão desatualizada }
  /homes/{homeId}/rooms:
    get:
      operationId: listRooms
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Ambientes da residência }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createRoom
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '201': { description: Ambiente criado }
  /homes/{homeId}/devices:
    get:
      operationId: listDevices
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Dispositivos normalizados da residência }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createDevice
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '201': { description: Dispositivo persistido }
  /homes/{homeId}/devices/{deviceId}:
    patch:
      operationId: updateDevice
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { name: deviceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Dispositivo atualizado }
        '404': { description: Dispositivo não encontrado na residência }
        '409': { $ref: '#/components/responses/VersionConflict' }
    delete:
      operationId: deleteDevice
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { name: deviceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: Dispositivo excluído }
        '409': { $ref: '#/components/responses/VersionConflict' }
  /homes/{homeId}/dashboard:
    get:
      operationId: getOperationalDashboard
      description: Agrega somente dados autorizados da residência para segurança, conectividade, dispositivos, câmeras, clima, energia, automações, atividade e sistema.
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Snapshot operacional gerado pelo backend }
        '403': { $ref: '#/components/responses/Forbidden' }
  /homes/{homeId}/floorplan:
    get:
      operationId: getFloorplan
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Planta e versão atuais }
    put:
      operationId: saveFloorplan
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Nova versão da planta persistida }
  /homes/{homeId}/scenes:
    get:
      operationId: listScenes
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Cenas da residência }
    post:
      operationId: createScene
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '201': { description: Cena criada }
  /homes/{homeId}/automations:
    get:
      operationId: listAutomations
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Automações da residência }
    post:
      operationId: createAutomation
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '201': { description: Automação criada }
  /homes/{homeId}/notifications:
    get:
      operationId: listNotifications
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Notificações da residência }
  /homes/{homeId}/energy:
    get:
      operationId: getEnergy
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Leituras, total, custo e configuração }
  /homes/{homeId}/assistant:
    post:
      operationId: executeAssistantCommand
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Comando interpretado dentro da residência }
  /homes/{homeId}/events:
    get:
      operationId: streamHomeEvents
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Stream SSE filtrado pela residência }
  /homes/{homeId}/integrations:
    get:
      operationId: listIntegrations
      description: Retorna somente metadados; credenciais cifradas nunca são expostas.
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/HomeId' }]
      responses:
        '200': { description: Integrações configuradas, status e última sincronização }
  /homes/{homeId}/integrations/{provider}/credentials:
    put:
      operationId: saveIntegrationCredentials
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/Provider' }
      responses:
        '200': { description: Credenciais cifradas e metadados retornados }
        '400': { description: Provedor ou credenciais inválidas }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: revokeIntegrationCredentials
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/Provider' }
      responses: { '204': { description: Credencial revogada } }
  /homes/{homeId}/integrations/{provider}/health:
    get:
      operationId: testIntegration
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/Provider' }
      responses:
        '200': { description: Provedor operacional e latência medida }
        '409': { description: Integração ainda não configurada }
        '502': { description: Falha isolada do provedor }
  /homes/{homeId}/integrations/{provider}/sync:
    post:
      operationId: syncIntegration
      description: Importa ou atualiza dispositivos por identificador externo e inicia a assinatura de eventos.
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/HomeId' }
        - { $ref: '#/components/parameters/Provider' }
      responses:
        '200': { description: Totais criados/atualizados e horário da sincronização }
        '409': { description: Integração ainda não configurada }
        '502': { description: Falha isolada do provedor }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ninho_access
  parameters:
    HomeId:
      name: homeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    FloorId:
      name: floorId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VersionQuery:
      name: version
      in: query
      required: true
      schema: { type: integer, minimum: 1 }
    Provider:
      name: provider
      in: path
      required: true
      schema: { type: string, enum: [tuya, home-assistant] }
  responses:
    Unauthorized:
      description: Sessão ausente ou inválida
    Forbidden:
      description: Usuário sem acesso à residência
    VersionConflict:
      description: A versão enviada está desatualizada; recarregue o recurso antes de tentar novamente