# server/routes/v1.js
import express from 'express';
import { AppError } from '../core/app-error.js';
import { authenticate, authorizeHome } from '../security/auth-middleware.js';
import { createHomeDomainRouter } from './home-domain.js';

const cookie = req => Object.fromEntries(String(req.get('cookie')||'').split(';').map(item=>item.trim().split('=').map(decodeURIComponent)).filter(pair=>pair.length===2));
const refreshCookie = {httpOnly:true,sameSite:'strict',secure:process.env.NODE_ENV==='production',path:'/api/v1/auth',maxAge:7*24*60*60*1000};
const accessCookie = {httpOnly:true,sameSite:'strict',secure:process.env.NODE_ENV==='production',path:'/',maxAge:15*60*1000};
const setSession=(res,result)=>{res.cookie('ninho_access',result.accessToken,accessCookie);res.cookie('ninho_refresh',result.refreshToken,refreshCookie);};
const cleanName=(value,label)=>{const name=String(value||'').trim();if(!name||name.length>80)throw new AppError('VALIDATION_ERROR',`${label} deve ter até 80 caracteres.`,400);return name;};
const resourceVersion=value=>{const version=Number(value);if(!Number.isInteger(version)||version<1)throw new AppError('VALIDATION_ERROR','Versão atual do recurso é obrigatória.',400);return version;};
const audit=(identity,req,type,targetId,metadata={})=>identity.record({type,homeId:req.params.homeId,actorId:req.auth.sub,targetId,result:'succeeded',correlationId:req.correlationId,metadata});

export function createV1Router({auth,identity,tokens,providers,vault,credentialStore=identity,turnstile,homeRepository,events,integrations,controlExternal,dashboard}) {
  const router=express.Router();const requireAuth=authenticate(tokens,identity);
  router.get('/health',(req,res)=>res.json({status:'ok',timestamp:new Date().toISOString(),correlationId:req.correlationId}));
  router.get('/auth/config',(_req,res)=>res.json(turnstile.publicConfig()));
  router.post('/auth/register',turnstile.middleware('register'),async(req,res,next)=>{try{const result=await auth.register(req.body);setSession(res,result);res.status(201).json({expiresIn:result.expiresIn,user:result.user})}catch(e){next(e)}});
  router.post('/auth/login',turnstile.middleware('login'),async(req,res,next)=>{try{const result=await auth.login(req.body?.email,req.body?.password);setSession(res,result);res.json({expiresIn:result.expiresIn,user:result.user})}catch(e){next(e)}});
  router.post('/auth/refresh',async(req,res,next)=>{try{const result=await auth.refresh(cookie(req).ninho_refresh);setSession(res,result);res.json({expiresIn:result.expiresIn,user:result.user})}catch(e){next(e)}});
  router.post('/auth/logout',async(req,res)=>{await auth.logout(cookie(req).ninho_refresh);res.clearCookie('ninho_access',{path:'/'});res.clearCookie('ninho_refresh',{path:'/api/v1/auth'});res.status(204).end()});
  router.post('/auth/password/forgot',turnstile.middleware('password-recovery'),async(req,res,next)=>{try{await auth.requestPasswordReset(req.body?.email);res.status(202).json({message:'Se o e-mail estiver cadastrado, enviaremos as instruções de recuperação.'})}catch(e){next(e)}});
  router.post('/auth/password/reset',turnstile.middleware('password-reset'),async(req,res,next)=>{try{await auth.resetPassword(req.body?.token,req.body?.password);res.status(204).end()}catch(e){next(e)}});
  router.get('/me',requireAuth,async(req,res)=>res.json(await identity.findUser(req.auth.sub)));
  router.get('/homes',requireAuth,async(req,res)=>res.json(await identity.listHomes(req.auth.sub)));
  router.post('/homes',requireAuth,async(req,res,next)=>{try{const name=String(req.body?.name||'').trim();if(!name)throw new AppError('VALIDATION_ERROR','Nome da residência é obrigatório.',400);const home=await identity.createHome({name,timezone:req.body?.timezone,ownerId:req.auth.sub});const floor=await identity.createFloor({homeId:home.id,name:'Térreo',position:0});for(const [position,room] of ['Sala','Quarto','Cozinha','Banheiro'].entries())await identity.createRoom({floorId:floor.id,name:room,position});await identity.record({type:'HOME_CREATED',homeId:home.id,actorId:req.auth.sub,targetId:home.id,result:'succeeded',correlationId:req.correlationId});res.status(201).json(home)}catch(e){next(e)}});
  router.patch('/homes/:homeId',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{const patch={version:resourceVersion(req.body?.version)};if(req.body?.name!==undefined)patch.name=cleanName(req.body.name,'Nome da residência');if(req.body?.timezone!==undefined){patch.timezone=String(req.body.timezone).trim();if(!patch.timezone||patch.timezone.length>80)throw new AppError('VALIDATION_ERROR','Fuso horário inválido.',400);}const home=await identity.updateHome(req.params.homeId,patch);if(home===false)throw new AppError('VERSION_CONFLICT','A residência foi alterada por outra sessão. Atualize e tente novamente.',409);if(!home)throw new AppError('NOT_FOUND','Residência não encontrada.',404);await audit(identity,req,'HOME_UPDATED',home.id,{version:home.version});res.json(home);}catch(e){next(e)}});
  router.delete('/homes/:homeId',requireAuth,authorizeHome(identity,['owner']),async(req,res,next)=>{try{const version=resourceVersion(req.query.version);const removed=await identity.deleteHome(req.params.homeId,req.auth.sub,version);if(removed==='last-home')throw new AppError('LAST_HOME','Crie outra residência antes de excluir esta.',409);if(removed===false)throw new AppError('VERSION_CONFLICT','A residência foi alterada por outra sessão.',409);if(!removed)throw new AppError('NOT_FOUND','Residência não encontrada.',404);res.status(204).end();}catch(e){next(e)}});
  router.get('/homes/:homeId/floors',requireAuth,authorizeHome(identity),async(req,res)=>res.json(await identity.listFloors(req.params.homeId)));
  router.post('/homes/:homeId/floors',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{const name=String(req.body?.name||'').trim();if(!name)throw new AppError('VALIDATION_ERROR','Nome do piso é obrigatório.',400);const current=await identity.listFloors(req.params.homeId);const floor=await identity.createFloor({homeId:req.params.homeId,name,position:req.body?.position??current.length});if(!floor)throw new AppError('FLOOR_ALREADY_EXISTS','Já existe um piso com esse nome.',409);await identity.record({type:'FLOOR_CREATED',homeId:req.params.homeId,actorId:req.auth.sub,targetId:floor.id,result:'succeeded',correlationId:req.correlationId});res.status(201).json(floor)}catch(e){next(e)}});
  router.patch('/homes/:homeId/floors/:floorId',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{const patch={version:resourceVersion(req.body?.version)};if(req.body?.name!==undefined)patch.name=cleanName(req.body.name,'Nome do piso');if(req.body?.position!==undefined){patch.position=Number(req.body.position);if(!Number.isInteger(patch.position)||patch.position<0)throw new AppError('VALIDATION_ERROR','Posição do piso inválida.',400);}const floor=await identity.updateFloor(req.params.homeId,req.params.floorId,patch);if(floor==='duplicate')throw new AppError('FLOOR_ALREADY_EXISTS','Já existe um piso com esse nome.',409);if(floor===false)throw new AppError('VERSION_CONFLICT','O piso foi alterado por outra sessão.',409);if(!floor)throw new AppError('NOT_FOUND','Piso não encontrado.',404);await audit(identity,req,'FLOOR_UPDATED',floor.id,{version:floor.version});res.json(floor);}catch(e){next(e)}});
  router.delete('/homes/:homeId/floors/:floorId',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{const removed=await identity.deleteFloor(req.params.homeId,req.params.floorId,resourceVersion(req.query.version));if(removed==='last-floor')throw new AppError('LAST_FLOOR','A residência precisa manter pelo menos um piso.',409);if(removed==='not-empty')throw new AppError('FLOOR_NOT_EMPTY','Exclua ou mova os cômodos antes de excluir o piso.',409);if(removed===false)throw new AppError('VERSION_CONFLICT','O piso foi alterado por outra sessão.',409);if(!removed)throw new AppError('NOT_FOUND','Piso não encontrado.',404);await audit(identity,req,'FLOOR_DELETED',req.params.floorId);res.status(204).end();}catch(e){next(e)}});
  router.get('/homes/:homeId/floors/:floorId/rooms',requireAuth,authorizeHome(identity),async(req,res,next)=>{try{if(!await identity.findFloor(req.params.homeId,req.params.floorId))throw new AppError('FLOOR_NOT_FOUND','Piso não encontrado.',404);res.json(await identity.listRooms(req.params.floorId))}catch(e){next(e)}});
  router.post('/homes/:homeId/floors/:floorId/rooms',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{if(!await identity.findFloor(req.params.homeId,req.params.floorId))throw new AppError('FLOOR_NOT_FOUND','Piso não encontrado.',404);const name=String(req.body?.name||'').trim();if(!name)throw new AppError('VALIDATION_ERROR','Nome do cômodo é obrigatório.',400);const current=await identity.listRooms(req.params.floorId);const room=await identity.createRoom({floorId:req.params.floorId,name,position:req.body?.position??current.length});if(!room)throw new AppError('ROOM_ALREADY_EXISTS','Já existe um cômodo com esse nome neste piso.',409);await identity.record({type:'ROOM_CREATED',homeId:req.params.homeId,actorId:req.auth.sub,targetId:room.id,result:'succeeded',correlationId:req.correlationId});res.status(201).json(room)}catch(e){next(e)}});
  router.get('/homes/:homeId/audit',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res)=>res.json(await identity.listAudit(req.params.homeId)));
  router.get('/homes/:homeId/integrations/health',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res)=>res.json(providers?await providers.health():{}));
  router.get('/homes/:homeId/integrations',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res)=>res.json(await credentialStore.listIntegrations(req.params.homeId)));
  router.get('/homes/:homeId/integrations/:provider/health',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{res.json(await integrations.health(req.params.homeId,String(req.params.provider)));}catch(e){next(e)}});
  router.post('/homes/:homeId/integrations/:provider/sync',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{res.json(await integrations.sync({homeId:req.params.homeId,providerName:String(req.params.provider),actorId:req.auth.sub,correlationId:req.correlationId}));}catch(e){next(e)}});
  router.put('/homes/:homeId/integrations/:provider/credentials',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res,next)=>{try{const provider=String(req.params.provider);if(!['tuya','home-assistant','openai'].includes(provider))throw new AppError('VALIDATION_ERROR','Provedor não suportado.',400);const credentials=req.body?.credentials;if(!credentials||typeof credentials!=='object'||Array.isArray(credentials)||!Object.keys(credentials).length)throw new AppError('VALIDATION_ERROR','Credenciais obrigatórias.',400);const sealed=vault.seal(credentials,{homeId:req.params.homeId,provider});const integration=await credentialStore.saveIntegrationCredential({homeId:req.params.homeId,provider,sealed,actorId:req.auth.sub});await identity.record({type:'INTEGRATION_CREDENTIALS_ROTATED',homeId:req.params.homeId,actorId:req.auth.sub,targetId:provider,result:'succeeded',correlationId:req.correlationId,metadata:{keyVersion:sealed.keyVersion}});res.json(integration)}catch(e){next(e)}});
  router.delete('/homes/:homeId/integrations/:provider/credentials',requireAuth,authorizeHome(identity,['owner','admin']),async(req,res)=>{const removed=await credentialStore.deleteIntegrationCredential(req.params.homeId,req.params.provider);await identity.record({type:'INTEGRATION_CREDENTIALS_REVOKED',homeId:req.params.homeId,actorId:req.auth.sub,targetId:req.params.provider,result:removed?'succeeded':'not_found',correlationId:req.correlationId});res.status(removed?204:404).end()});
  if(homeRepository)router.use('/homes/:homeId',requireAuth,authorizeHome(identity),createHomeDomainRouter({identity,repository:homeRepository,events,controlExternal,dashboard}));
  return router;
}